+++ routing/http/server/server.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"mime"
 	"net/http"
-	"strconv"
 	"strings"
 	"time"
 
@@ -23,6 +22,7 @@ import (
 	"github.com/ipfs/go-cid"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/multiformats/go-multiaddr"
+	"github.com/multiformats/go-multibase"
 
 	logging "github.com/ipfs/go-log/v2"
 )
@@ -242,16 +242,21 @@ func (s *server) findProvidersNDJSON(w http.ResponseWriter, provIter iter.Result
 func (s *server) findPeers(w http.ResponseWriter, r *http.Request) {
 	pidStr := mux.Vars(r)["peer-id"]
 
-	// pidStr must be in CIDv1 format. Therefore, use [cid.Decode]. We can't use
-	// [peer.Decode] because that would allow other formats to pass through.
+	// While specification states that peer-id is expected to be in CIDv1 format, reality
+	// is the clients will often learn legacy PeerID string from other sources,
+	// and try to use it.
+	// See https://github.com/libp2p/specs/blob/master/peer-ids/peer-ids.md#string-representation
+	// We are liberal in inputs here, and uplift legacy PeerID to CID if necessary.
+	// Rationale: it is better to fix this common mistake than to error and break peer routing.
 	cid, err := cid.Decode(pidStr)
 	if err != nil {
-		if pid, err := peer.Decode(pidStr); err == nil {
-			writeErr(w, "FindPeers", http.StatusBadRequest, fmt.Errorf("the value is a peer ID, try using its CID representation: %s", peer.ToCid(pid).String()))
+		// check if input is peer ID in legacy format
+		if pid, err2 := peer.Decode(pidStr); err2 == nil {
+			cid = peer.ToCid(pid)
 		} else {
-			writeErr(w, "FindPeers", http.StatusBadRequest, fmt.Errorf("unable to parse peer ID: %w", err))
+			writeErr(w, "FindPeers", http.StatusBadRequest, fmt.Errorf("unable to parse peer ID as libp2p-key CID: %w", err))
+			return
 		}
-		return
 	}
 
 	pid, err := peer.FromCid(cid)
@@ -366,8 +371,14 @@ func (s *server) findPeersNDJSON(w http.ResponseWriter, peersIter iter.ResultIte
 }
 
 func (s *server) GetIPNS(w http.ResponseWriter, r *http.Request) {
-	if !strings.Contains(r.Header.Get("Accept"), mediaTypeIPNSRecord) {
-		writeErr(w, "GetIPNS", http.StatusNotAcceptable, errors.New("content type in 'Accept' header is missing or not supported"))
+	acceptHdrValue := r.Header.Get("Accept")
+	// When 'Accept' header is missing, default to 'application/vnd.ipfs.ipns-record'
+	// (improved UX, similar to how we default to JSON response for /providers and /peers)
+	if len(acceptHdrValue) == 0 || strings.Contains(acceptHdrValue, mediaTypeWildcard) {
+		acceptHdrValue = mediaTypeIPNSRecord
+	}
+	if !strings.Contains(acceptHdrValue, mediaTypeIPNSRecord) {
+		writeErr(w, "GetIPNS", http.StatusNotAcceptable, errors.New("content type in 'Accept' header is not supported, retry with 'application/vnd.ipfs.ipns-record'"))
 		return
 	}
 
@@ -397,15 +408,31 @@ func (s *server) GetIPNS(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if ttl, err := record.TTL(); err == nil {
-		w.Header().Set("Cache-Control", fmt.Sprintf("max-age=%d", int(ttl.Seconds())))
+	var remainingValidity int
+	// Include 'Expires' header with time when signature expiration happens
+	if validityType, err := record.ValidityType(); err == nil && validityType == ipns.ValidityEOL {
+		if validity, err := record.Validity(); err == nil {
+			w.Header().Set("Expires", validity.UTC().Format(http.TimeFormat))
+			remainingValidity = int(time.Until(validity).Seconds())
+		}
 	} else {
-		w.Header().Set("Cache-Control", "max-age=60")
+		remainingValidity = int(ipns.DefaultRecordLifetime.Seconds())
 	}
+	if ttl, err := record.TTL(); err == nil {
+		setCacheControl(w, int(ttl.Seconds()), remainingValidity)
+	} else {
+		setCacheControl(w, int(ipns.DefaultRecordTTL.Seconds()), remainingValidity)
+	}
+	w.Header().Set("Last-Modified", time.Now().UTC().Format(http.TimeFormat))
 
-	recordEtag := strconv.FormatUint(xxhash.Sum64(rawRecord), 32)
-	w.Header().Set("Etag", recordEtag)
+	w.Header().Set("Etag", fmt.Sprintf(`"%x"`, xxhash.Sum64(rawRecord)))
 	w.Header().Set("Content-Type", mediaTypeIPNSRecord)
+
+	// Content-Disposition is not required, but improves UX by assigning a meaningful filename when opening URL in a web browser
+	if filename, err := cid.StringOfBase(multibase.Base36); err == nil {
+		w.Header().Set("Content-Disposition", fmt.Sprintf("attachment; filename=\"%s.ipns-record\"", filename))
+	}
+	w.Header().Add("Vary", "Accept")
 	w.Write(rawRecord)
 }
 
@@ -457,8 +484,30 @@ func (s *server) PutIPNS(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 }
 
-func writeJSONResult(w http.ResponseWriter, method string, val any) {
+var (
+	// Rule-of-thumb Cache-Control policy is to work well with caching proxies and load balancers.
+	// If there are any results, cache on the client for longer, and hint any in-between caches to
+	// serve cached result and upddate cache in background as long we have
+	// result that is within Amino DHT expiration window
+	maxAgeWithResults    = int((5 * time.Minute).Seconds())  // cache >0 results for longer
+	maxAgeWithoutResults = int((15 * time.Second).Seconds()) // cache no results briefly
+	maxStale             = int((48 * time.Hour).Seconds())   // allow stale results as long within Amino DHT  Expiration window
+)
+
+func setCacheControl(w http.ResponseWriter, maxAge int, stale int) {
+	w.Header().Set("Cache-Control", fmt.Sprintf("public, max-age=%d, stale-while-revalidate=%d, stale-if-error=%d", maxAge, stale, stale))
+}
+
+func writeJSONResult(w http.ResponseWriter, method string, val interface{ Length() int }) {
 	w.Header().Add("Content-Type", mediaTypeJSON)
+	w.Header().Add("Vary", "Accept")
+
+	if val.Length() > 0 {
+		setCacheControl(w, maxAgeWithResults, maxStale)
+	} else {
+		setCacheControl(w, maxAgeWithoutResults, maxStale)
+	}
+	w.Header().Set("Last-Modified", time.Now().UTC().Format(http.TimeFormat))
 
 	// keep the marshaling separate from the writing, so we can distinguish bugs (which surface as 500)
 	// from transient network issues (which surface as transport errors)
@@ -495,19 +544,28 @@ func writeResultsIterNDJSON[T any](w http.ResponseWriter, resultIter iter.Result
 	defer resultIter.Close()
 
 	w.Header().Set("Content-Type", mediaTypeNDJSON)
-	w.WriteHeader(http.StatusOK)
+	w.Header().Add("Vary", "Accept")
+	w.Header().Set("Last-Modified", time.Now().UTC().Format(http.TimeFormat))
 
+	hasResults := false
 	for resultIter.Next() {
 		res := resultIter.Val()
 		if res.Err != nil {
 			logger.Errorw("ndjson iterator error", "Error", res.Err)
 			return
 		}
+
 		// don't use an encoder because we can't easily differentiate writer errors from encoding errors
 		b, err := drjson.MarshalJSONBytes(res.Val)
 		if err != nil {
 			logger.Errorw("ndjson marshal error", "Error", err)
 			return
+		}
+
+		if !hasResults {
+			hasResults = true
+			// There's results, cache useful result for longer
+			setCacheControl(w, maxAgeWithResults, maxStale)
 		}
 
 		_, err = w.Write(b)
@@ -525,5 +583,10 @@ func writeResultsIterNDJSON[T any](w http.ResponseWriter, resultIter iter.Result
 		if f, ok := w.(http.Flusher); ok {
 			f.Flush()
 		}
+	}
+
+	if !hasResults {
+		// There weren't results, cache for shorter
+		setCacheControl(w, maxAgeWithoutResults, maxStale)
 	}
 }